ingestion/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
scikit-learn==1.6.1
numpy==2.2.0
python-dotenv==1.0.1
